+++ bamt/networks/base.py
@@ -27,7 +27,8 @@ from bamt.log import logger_network
 from bamt.nodes.base import BaseNode
 from bamt.utils import graph_utils, serialization_utils, check_utils
 
-from bamt.utils.check_utils import NetworkChecker
+from bamt.checkers.network_checker import NetworkChecker
+from bamt.checkers.enums import nodes_with_regressors
 
 
 class BaseNetwork(object):
@@ -287,6 +288,8 @@ class BaseNetwork(object):
 
     def set_edges(self, edges: Optional[List[Sequence[str]]] = None):
         """
+        NB!: this method affects only on attributes and can't define structure (with 2 stage)
+
         additional function to set edges manually. User should be aware that
         nodes must be a subclass of BaseNode.
         param: edges dict with name and node (if a lot of nodes should be added)
@@ -300,8 +303,8 @@ class BaseNetwork(object):
                 if self[node1] and self[node2]:
                     if (
                         not self.has_logit
-                        and self.checker.checker_descriptor["types"][node1].is_cont
-                        and self.checker.checker_descriptor["types"][node2].is_disc
+                        and self.checker[node1]["node_checker"].is_cont
+                        and self.checker[node2]["node_checker"].is_disc
                     ):
                         logger_network.warning(
                             f"Restricted edge detected (has_logit=False) : [{node1}, {node2}]"
@@ -331,16 +334,18 @@ class BaseNetwork(object):
         nodes, edges:
         """
         if nodes and (info or (self.descriptor["types"] and self.descriptor["signs"])):
-            if (
-                any("mixture" in node.type.lower() for node in nodes)
-                and not self.use_mixture
-            ):
+            # if (
+            #     any("mixture" in node.type.lower() for node in nodes)
+            #     and not self.use_mixture
+            # ):
+            if self.checker.has_mixture_nodes() and not self.use_mixture:
                 logger_network.error("Compatibility error: use mixture.")
                 return
-            if (
-                any("logit" in node.type.lower() for node in nodes)
-                and not self.has_logit
-            ):
+            # if (
+            #     any("logit" in node.type.lower() for node in nodes)
+            #     and not self.has_logit
+            # ):
+            if self.checker.has_logit_nodes() and not self.has_logit:
                 logger_network.error("Compatibility error: has logit.")
                 return
 
@@ -350,25 +355,35 @@ class BaseNetwork(object):
                 logger_network.error("Nodes/info detection failed.")
                 return
 
-            builder = builders.builders_base.VerticesDefiner(
-                descriptor=self.descriptor, regressor=None
+            vertices_builder = builders.builders_base.VerticesDefiner(
+                descriptor=self.descriptor,
+                checkers_rules=self.checker.get_checker_rules(),
+                regressor=None,
             )  # init worker
 
-            self.nodes = builder.init_nodes() # 1 stage
+            self.nodes = vertices_builder.init_nodes()  # 1 stage
             if len(edges) != 0:
                 # set edges and register members
                 self.set_edges(edges=edges)
-                builder.structure = self.edges
-                builder.get_family()
 
-            builder.overwrite_vertex(
+                vertices_builder.structure = self.edges
+                vertices_builder.vertices = self.nodes
+                vertices_builder.get_family()
+
+            edges_builder = builders.builders_base.EdgesDefiner(
+                vertices=self.nodes,
+                descriptor=self.descriptor,
+                checkers_rules=self.checker.get_checker_rules(),
+            )
+
+            edges_builder.overwrite_vertex(
                 has_logit=self.has_logit,
                 use_mixture=self.use_mixture,
                 classifier=None,
                 regressor=None,
             )
 
-            self.set_nodes(nodes=builder.structure)
+            self.set_nodes(nodes=edges_builder.vertices)
 
     def _param_validation(self, params: Dict[str, Any]) -> bool:
         if all(self[i] for i in params.keys()):
@@ -450,48 +465,32 @@ class BaseNetwork(object):
             json.dump(data, out)
         return True
 
-    def save_params(self, outdir: str):
-        """
-        Function to save BN params to json file
-        outdir: output directory
-        """
-        return self._save_to_file(outdir, self.distributions)
-
-    def save_structure(self, outdir: str):
-        """
-        Function to save BN edges to json file
-        outdir: output directory
-        """
-        return self._save_to_file(outdir, self.edges)
-
-    def save(self, bn_name, models_dir: str = "models_dir"):
-        """
-        Function to save the whole BN to json file.
-
-        :param bn_name: unique name of bn user want to save. It will be used as file name (e.g. bn_name.json).
-        :param models_dir: if picklization is broken, joblib will serialize models in compressed files
-        in models directory.
-
-        :return: saving status.
-        """
+    def _serialize_dist(self, bn_name, models_dir):
         distributions = deepcopy(self.distributions)
-        new_weights = {str(key): self.weights[key] for key in self.weights}
-
         to_serialize = {}
         # separate logit and gaussian nodes from distributions to serialize bn's models
         for node_name in distributions.keys():
-            if "Mixture" in self[node_name].type:
+            node_checker = self.checker[node_name]["node_checker"]
+            # if "Mixture" in self[node_name].type:
+            if node_checker.is_mixture:
                 continue
-            if self[node_name].type.startswith("Gaussian"):
-                if not distributions[node_name]["regressor"]:
-                    continue
+            # if self[node_name].type.startswith("Gaussian"):
+            # if node_checker.does_require_regressor and not node_checker.has_combinations: ???
+            #     if not distributions[node_name]["regressor"]:
+            #         continue
+            if node_checker.root:
+                continue
+            # if (
+            #     "Gaussian" in self[node_name].type
+            #     or "Logit" in self[node_name].type
+            #     or "ConditionalLogit" in self[node_name].type
+            # ):
             if (
-                "Gaussian" in self[node_name].type
-                or "Logit" in self[node_name].type
-                or "ConditionalLogit" in self[node_name].type
+                node_checker.does_require_classifier
+                or node_checker.does_require_regressor
             ):
                 to_serialize[node_name] = [
-                    self[node_name].type,
+                    node_checker,
                     distributions[node_name],
                 ]
 
@@ -502,6 +501,36 @@ class BaseNetwork(object):
 
         for serialized_node in serialized_dist.keys():
             distributions[serialized_node] = serialized_dist[serialized_node]
+        return distributions
+
+    def save_params(self, outdir: str, bn_name, models_dir: str = "models_dir"):
+        """
+        Function to save BN params to json file
+        outdir: output directory
+        """
+
+        return self._save_to_file(outdir, self._serialize_dist(bn_name, models_dir))
+
+    def save_structure(self, outdir: str):
+        """
+        Function to save BN edges to json file
+        outdir: output directory
+        """
+        return self._save_to_file(outdir, self.edges)
+
+    def save(self, outdir, bn_name, models_dir: str = "models_dir"):
+        """
+        Function to save the whole BN to json file.
+
+        :param bn_name: unique name of bn user want to save. It will be used as file name (e.g. bn_name.json).
+        :param models_dir: if picklization is broken, joblib will serialize models in compressed files
+        in models directory.
+
+        :return: saving status.
+        """
+
+        new_weights = {str(key): self.weights[key] for key in self.weights}
+        distributions = self._serialize_dist(bn_name, models_dir)
 
         outdict = {
             "info": self.descriptor,
@@ -509,11 +538,9 @@ class BaseNetwork(object):
             "parameters": distributions,
             "weights": new_weights,
         }
-        return self._save_to_file(f"{bn_name}.json", outdict)
+        return self._save_to_file(outdir, outdict)
 
-    def load(self,
-             input_data: Union[str, Dict],
-             models_dir: str = "/"):
+    def load(self, input_data: Union[str, Dict], models_dir: str = "/"):
         """
         Function to load the whole BN from json file.
         :param input_data: input path to json file with bn.
@@ -533,34 +560,39 @@ class BaseNetwork(object):
         self.add_nodes(input_dict["info"])
         self.set_structure(edges=input_dict["edges"])
 
-        # check compatibility with father network.
-        if not self.use_mixture:
-            for node_data in input_dict["parameters"].values():
-                if "hybcprob" not in node_data.keys():
-                    continue
-                else:
-                    # Since we don't have information about types of nodes, we
-                    # should derive it from parameters.
-                    if any(
-                        list(node_keys.keys()) == ["covars", "mean", "coef"]
-                        for node_keys in node_data["hybcprob"].values()
-                    ):
-                        logger_network.error(
-                            f"This crucial parameter is not the same as father's parameter: use_mixture."
-                        )
-                        return
+        validation_result = self.checker.validate_load(input_dict, self)
+        if isinstance(validation_result, str):
+            logger_network.error(f"This crucial parameter is not the same as father's parameter: {validation_result}.")
+            return
 
-        # check if edges before and after are the same.They can be different in
-        # the case when user sets forbidden edges.
-        if not self.has_logit:
-            if not all(
-                edges_before == [edges_after[0], edges_after[1]]
-                for edges_before, edges_after in zip(input_dict["edges"], self.edges)
-            ):
-                logger_network.error(
-                    f"This crucial parameter is not the same as father's parameter: has_logit."
-                )
-                return
+        # # check compatibility with father network.
+        # if not self.use_mixture:
+        #     for node_data in input_dict["parameters"].values():
+        #         if "hybcprob" not in node_data.keys():
+        #             continue
+        #         else:
+        #             # Since we don't have information about types of nodes, we
+        #             # should derive it from parameters.
+        #             if any(
+        #                 list(node_keys.keys()) == ["covars", "mean", "coef"]
+        #                 for node_keys in node_data["hybcprob"].values()
+        #             ):
+        #                 logger_network.error(
+        #                     f"This crucial parameter is not the same as father's parameter: use_mixture."
+        #                 )
+        #                 return
+        #
+        # # check if edges before and after are the same.They can be different in
+        # # the case when user sets forbidden edges.
+        # if not self.has_logit:
+        #     if not all(
+        #         edges_before == [edges_after[0], edges_after[1]]
+        #         for edges_before, edges_after in zip(input_dict["edges"], self.edges)
+        #     ):
+        #         logger_network.error(
+        #             f"This crucial parameter is not the same as father's parameter: has_logit."
+        #         )
+        #         return
 
         deserializer = serialization_utils.Deserializer(models_dir)
